Firmware hotfix, v0.0
Hotfix for the firmware for Longbow widget. EEPROM was randomly being overwritten/altered, removed all write to EEPROM commands to provide stability. Removed ability to set address/baud rate on the fly, need to find origin of issue and real fix. I think it has to do with the power on condition and things being written to randomly. Firmware deployed, official v0.0